可查询当前话费余额

+++ ZKLibs/sms.h
@@ -13,7 +13,7 @@
 typedef struct {
 	uint8_t numberType;
 	uint8_t encodeType;
-	uint8_t contentLen;
+	uint16_t contentLen;
 	uint8_t number[15];
 	int8_t time[15];
 	int8_t content[700];
